Refactor client routing with React Router
