+++ client/src/components/PartySetup.module.css
@@ -1,4 +1,4 @@
-/* client/src/components/PartySetupScreen.module.css */
+/* client/src/components/PartySetup.module.css */
 .screen {
   padding: 20px;
   font-family: Arial, sans-serif;
