+++ client/src/main.jsx
@@ -1,6 +1,5 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter } from 'react-router-dom'
 import './index.css'
 import App from './App.jsx'
 import { useGameStore } from './store/gameStore'
@@ -12,12 +11,10 @@ useGameStore.getState().load()
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <BrowserRouter>
-      <ModalProvider>
-        <NotificationProvider>
-          <App />
-        </NotificationProvider>
-      </ModalProvider>
-    </BrowserRouter>
+    <ModalProvider>
+      <NotificationProvider>
+        <App />
+      </NotificationProvider>
+    </ModalProvider>
   </StrictMode>,
 )
